+++ labs/CultBook/model/Livro.cs
@@ -7,11 +7,11 @@ public abstract class Livro
     public string Descricao { get; set; }
     public string Autor { get; set; }
     public int Estoque { get; set; }
-    public double Preco { get; set; }
+    public decimal Preco { get; set; }
     public string Categoria { get; set; }
 
     protected Livro(string isbn, string titulo, string descricao, string autor,
-                         int estoque, double preco, string categoria)
+                         int estoque, decimal preco, string categoria)
     {
         Isbn = isbn;
         Titulo = titulo;
@@ -22,7 +22,7 @@ public abstract class Livro
         Categoria = categoria;
     }
 
-    public abstract double CalcularPrecoTotal();
+    public abstract decimal CalcularPrecoTotal();
 
     public override string ToString()
     {
